+++ src/mcmc/abstractmcmc.jl
@@ -69,6 +69,18 @@ function find_initial_params_ldf(
     )
 end
 
+"""
+    post_sample_hook(chain, sampler::AbstractSampler; kwargs...)
+
+A post-sampling hook that can e.g. print info about the results of sampling.
+
+Implementations of this should be careful to take `kwargs...` as keyword arguments instead
+of restricting the signature to specific keyword arguments. Right now, the only keyword
+argument that is passed to this function is `verbose`, but in the future additional keyword
+arguments may be passed here.
+"""
+post_sample_hook(chain, ::AbstractSampler; kwargs...) = nothing
+
 #########################################
 # Default definitions for the interface #
 #########################################
@@ -87,18 +99,22 @@ function AbstractMCMC.sample(
     initial_params=init_strategy(spl),
     check_model::Bool=true,
     chain_type=DEFAULT_CHAIN_TYPE,
+    verbose::Bool=true,
     kwargs...,
 )
     check_model && Turing._check_model(model, spl)
-    return AbstractMCMC.mcmcsample(
+    chain = AbstractMCMC.mcmcsample(
         rng,
         model,
         spl,
         N;
         initial_params=Turing._convert_initial_params(initial_params),
         chain_type,
+        verbose,
         kwargs...,
     )
+    post_sample_hook(chain, spl; verbose)
+    return chain
 end
 
 function AbstractMCMC.sample(
@@ -123,6 +139,7 @@ function AbstractMCMC.sample(
     n_chains::Integer;
     chain_type=DEFAULT_CHAIN_TYPE,
     check_model::Bool=true,
+    verbose::Bool=true,
     initial_params=fill(init_strategy(spl), n_chains),
     kwargs...,
 )
@@ -131,7 +148,7 @@ function AbstractMCMC.sample(
         errmsg = "`initial_params` must be an AbstractVector of length `n_chains`; one element per chain"
         throw(ArgumentError(errmsg))
     end
-    return AbstractMCMC.mcmcsample(
+    chain = AbstractMCMC.mcmcsample(
         rng,
         model,
         spl,
@@ -141,27 +158,11 @@ function AbstractMCMC.sample(
         chain_type,
         check_model=false, # no need to check again
         initial_params=map(Turing._convert_initial_params, initial_params),
+        verbose,
         kwargs...,
     )
-end
-
-"""
-    loadstate(chain::MCMCChains.Chains)
-
-Load the final state of the sampler from a `MCMCChains.Chains` object.
-
-To save the final state of the sampler, you must use `sample(...; save_state=true)`. If this
-argument was not used during sampling, calling `loadstate` will throw an error.
-"""
-function loadstate(chain::MCMCChains.Chains)
-    if !haskey(chain.info, :samplerstate)
-        throw(
-            ArgumentError(
-                "the chain object does not contain the final state of the sampler; to save the final state you must sample with `save_state=true`",
-            ),
-        )
-    end
-    return chain.info[:samplerstate]
+    post_sample_hook(chain, spl; verbose)
+    return chain
 end
 
 # TODO(penelopeysm): Remove initialstep and generalise MCMC sampling procedures