+++ test/stdlib/distributions.jl
@@ -1,6 +1,5 @@
 module DistributionsTests
 
-using ..NumericalTests: check_dist_numerical
 using Distributions
 using LinearAlgebra: I
 using Random: Random
@@ -9,11 +8,37 @@ using StatsFuns: logistic
 using Test: @testset, @test
 using Turing
 
+function check_dist_numerical(
+    dist, chn; mean_atol=0.1, mean_rtol=0.1, var_atol=1.0, var_rtol=0.5
+)
+    @testset "numerical" begin
+        # Extract values.
+        chn_xs = chn[@varname(x)]
+
+        # Check means.
+        dist_mean = mean(dist)
+        if !all(isnan, dist_mean) && !all(isinf, dist_mean)
+            chn_mean = mean(chn_xs)
+            @test chn_mean ≈ dist_mean atol = mean_atol rtol = mean_rtol
+        end
+
+        # Check variances.
+        # var() for Distributions.MatrixDistribution is not defined
+        if !(dist isa Distributions.MatrixDistribution)
+            # Variance
+            dist_var = var(dist)
+            if !all(isnan, dist_var) && !all(isinf, dist_var)
+                chn_var = var(chn_xs)
+                @test chn_var ≈ chn_var atol = var_atol rtol = var_rtol
+            end
+        end
+    end
+end
+
 @testset "distributions.jl" begin
-    rng = StableRNG(12345)
     @testset "distributions functions" begin
         ns = 10
-        logitp = randn(rng)
+        logitp = randn()
         d1 = BinomialLogit(ns, logitp)
         d2 = Binomial(ns, logistic(logitp))
         k = 3
@@ -24,7 +49,7 @@ using Turing
         d = OrderedLogistic(-2, [-1, 1])
 
         n = 1_000_000
-        y = rand(rng, d, n)
+        y = rand(d, n)
         K = length(d.cutpoints) + 1
         p = [mean(==(k), y) for k in 1:K]          # empirical probs
         pmf = [exp(logpdf(d, k)) for k in 1:K]
@@ -52,9 +77,10 @@ using Turing
 
     @testset "single distribution correctness" begin
         n_samples = 10_000
-        mean_tol = 0.1
+        mean_atol = 0.1
+        mean_rtol = 0.1
         var_atol = 1.0
-        var_tol = 0.5
+        var_rtol = 0.5
         multi_dim = 4
         # 1. UnivariateDistribution
         # NOTE: Noncentral distributions are commented out because of
@@ -130,22 +156,22 @@ using Turing
 
                             @model m() = x ~ dist
 
-                            seed = if dist isa GeneralizedExtremeValue
-                                # GEV is prone to giving really wacky results that are quite
-                                # seed-dependent.
-                                StableRNG(469)
-                            else
-                                StableRNG(468)
-                            end
-                            chn = sample(seed, m(), HMC(0.05, 20), n_samples)
+                            # Note: GeneralizedExtremeValue is prone to giving really wacky
+                            # results that are quite seed-dependent. Do not hesitate to
+                            # change the seed if the test fails, even by a large margin.
+                            seed = StableRNG(468)
+                            chn = sample(
+                                seed, m(), HMC(0.05, 20), n_samples; progress=false
+                            )
 
                             # Numerical tests.
                             check_dist_numerical(
                                 dist,
                                 chn;
-                                mean_tol=mean_tol,
+                                mean_atol=mean_atol,
+                                mean_rtol=mean_rtol,
                                 var_atol=var_atol,
-                                var_tol=var_tol,
+                                var_rtol=var_rtol,
                             )
                         end
                     end