+++ HISTORY.md
@@ -1,3 +1,25 @@
+# 0.45.0
+
+## Breaking changes
+
+Make FlexiChains the default chain type for MCMC sampling.
+
+MCMCChains is still fully supported: you can specify `chain_type=MCMCChains.Chains` in the `sample` function to use it instead.
+However, it is no longer loaded as a dependency of Turing and re-exported (it is now an extension).
+That means that if you were previously importing MCMCChains via Turing, you will now have to import it directly.
+
+# 0.44.5
+
+Allow users to disable the post-sample hook by passing `verbose=false` keyword argument to `sample`.
+
+# 0.44.4
+
+Add post-sampling warning message when there are divergent transitions with `HMC`, `NUTS` or `HMCDA`.
+
+# 0.44.3
+
+Add compatibility with SciMLBase v3.
+
 # 0.44.2
 
 Fix a bug in v0.44 where extra keyword arguments passed to `vi` (e.g. `callback`) would cause Turing to error.